Delt serviceutils inn i flere forskjellige filer
