+++ apps/testnorge-arena/arena-core/src/main/java/no/nav/registre/arena/core/service/util/ArbeidssoekerUtils.java
@@ -1,0 +1,134 @@
+package no.nav.registre.arena.core.service.util;
+
+import com.fasterxml.jackson.core.type.TypeReference;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.google.common.io.Resources;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import no.nav.registre.arena.core.consumer.rs.request.RettighetRequest;
+import no.nav.registre.arena.core.service.BrukereService;
+import no.nav.registre.arena.core.service.exception.ArbeidssoekerException;
+import no.nav.registre.testnorge.domain.dto.arena.testnorge.brukere.Kvalifiseringsgrupper;
+import org.springframework.stereotype.Service;
+
+import java.io.IOException;
+import java.net.URL;
+
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Random;
+
+import java.util.stream.Collectors;
+
+@Slf4j
+@Service
+@RequiredArgsConstructor
+public class ArbeidssoekerUtils {
+
+    private static final Map<String, List<KodeMedSannsynlighet>> aktivitestsfaserMedInnsats;
+
+    private final Random rand;
+    private final BrukereService brukereService;
+    private final ServiceUtils serviceUtils;
+
+    static {
+        aktivitestsfaserMedInnsats = new HashMap<>();
+        URL resourceInnsatser = Resources.getResource("aktfase_til_innsats.json");
+        ObjectMapper objectMapper = new ObjectMapper();
+        try {
+            Map<String, List<KodeMedSannsynlighet>> map = objectMapper.readValue(resourceInnsatser, new TypeReference<>() {
+            });
+            aktivitestsfaserMedInnsats.putAll(map);
+        } catch (IOException e) {
+            log.error("Kunne ikke laste inn innsatskoder.", e);
+        }
+    }
+
+    public List<RettighetRequest> opprettArbeidssoekerAap(
+            List<RettighetRequest> rettigheter,
+            String miljoe
+    ) {
+        return opprettArbeidssoeker(rettigheter, miljoe, rand.nextBoolean() ? Kvalifiseringsgrupper.BATT : Kvalifiseringsgrupper.VARIG);
+    }
+
+    public List<RettighetRequest> opprettArbeidssoekerAap(
+            List<RettighetRequest> rettigheter,
+            String miljoe,
+            String aktivitetsfase
+    ) {
+        if (aktivitetsfase == null || aktivitetsfase.isBlank()) {
+            return opprettArbeidssoeker(rettigheter, miljoe, Kvalifiseringsgrupper.BATT);
+        } else {
+            return opprettArbeidssoeker(rettigheter, miljoe, velgKvalifiseringsgruppeBasertPaaAktivitetsfase(aktivitetsfase));
+        }
+    }
+
+    public List<RettighetRequest> opprettArbeidssoekerTiltak(
+            List<RettighetRequest> rettigheter,
+            String miljoe
+    ) {
+        return opprettArbeidssoeker(rettigheter, miljoe, rand.nextBoolean() ? Kvalifiseringsgrupper.BATT : Kvalifiseringsgrupper.BFORM);
+    }
+
+    public List<RettighetRequest> opprettArbeidssoekerTillegg(
+            List<RettighetRequest> rettigheter,
+            String miljoe
+    ) {
+        return opprettArbeidssoeker(rettigheter, miljoe, rand.nextBoolean() ? Kvalifiseringsgrupper.BATT : Kvalifiseringsgrupper.BFORM);
+    }
+
+    private List<RettighetRequest> opprettArbeidssoeker(
+            List<RettighetRequest> rettigheter,
+            String miljoe,
+            Kvalifiseringsgrupper kvalifiseringsgruppe
+    ) {
+        var identerIArena = brukereService.hentEksisterendeArbeidsoekerIdenter();
+        var uregistrerteBrukere = rettigheter.stream().filter(rettighet -> !identerIArena.contains(rettighet.getPersonident())).map(RettighetRequest::getPersonident)
+                .collect(Collectors.toSet());
+
+        if (!uregistrerteBrukere.isEmpty()) {
+            var nyeBrukereResponse = brukereService
+                    .sendArbeidssoekereTilArenaForvalter(new ArrayList<>(uregistrerteBrukere), miljoe, kvalifiseringsgruppe);
+            List<String> feiledeIdenter = new ArrayList<>();
+            if (nyeBrukereResponse != null && nyeBrukereResponse.getNyBrukerFeilList() != null && !nyeBrukereResponse.getNyBrukerFeilList().isEmpty()) {
+                nyeBrukereResponse.getNyBrukerFeilList().forEach(nyBrukerFeil -> {
+                    log.error("Kunne ikke opprette ny bruker med fnr {} i arena: {}", nyBrukerFeil.getPersonident(), nyBrukerFeil.getMelding());
+                    feiledeIdenter.add(nyBrukerFeil.getPersonident());
+                });
+            }
+            rettigheter.removeIf(rettighet -> feiledeIdenter.contains(rettighet.getPersonident()));
+        }
+        return rettigheter;
+    }
+
+    public void opprettArbeidssoekerTiltakdeltakelse(
+            String ident,
+            String miljoe
+    ) {
+        var kvalifiseringsgruppe = rand.nextBoolean() ? Kvalifiseringsgrupper.BATT : Kvalifiseringsgrupper.BFORM;
+        var identerIArena = brukereService.hentEksisterendeArbeidsoekerIdenter();
+        var uregistrertBruker = !identerIArena.contains(ident);
+
+        if (uregistrertBruker) {
+            var nyeBrukereResponse = brukereService
+                    .sendArbeidssoekereTilArenaForvalter(Collections.singletonList(ident), miljoe, kvalifiseringsgruppe);
+            if (nyeBrukereResponse != null && nyeBrukereResponse.getNyBrukerFeilList() != null && !nyeBrukereResponse.getNyBrukerFeilList().isEmpty()) {
+                nyeBrukereResponse.getNyBrukerFeilList().forEach(nyBrukerFeil ->
+                        log.error("Kunne ikke opprette ny bruker med fnr {} i arena: {}", nyBrukerFeil.getPersonident(), nyBrukerFeil.getMelding())
+                );
+            }
+        }
+    }
+
+    private Kvalifiseringsgrupper velgKvalifiseringsgruppeBasertPaaAktivitetsfase(String aktivitetsfase) {
+        if (aktivitestsfaserMedInnsats.containsKey(aktivitetsfase)) {
+            var innsats = serviceUtils.velgKodeBasertPaaSannsynlighet(aktivitestsfaserMedInnsats.get(aktivitetsfase)).getKode();
+            return Kvalifiseringsgrupper.valueOf(innsats);
+        } else {
+            throw new ArbeidssoekerException("Ukjent aktivitetsfase " + aktivitetsfase);
+        }
+    }
+}
